feat: Dot 컴포넌트 상태별 bg-color 단일 책임 적용 및 디렉토리 구조 수정 (#30)
