+++ src/entities/ui/atoms/dot/index.tsx
@@ -1,23 +1,18 @@
 import { View, ViewProps } from "react-native";
 
 interface Props extends Omit<ViewProps, "children"> {
-  status: "available" | "busy" | "full";
+  status: DotStatus;
 }
 
 export default function Dot({ status, ...rest }: Props) {
-  let background = "";
-
-  switch (status) {
-    case "available":
-      background = "bg-green-200";
-      break;
-    case "busy":
-      background = "bg-orange-200";
-      break;
-    case "full":
-      background = "bg-red-200";
-      break;
-  }
+  const background = DotVariant[status];
 
   return <View className={`h-1 w-1 rounded-full ${background}`} {...rest} />;
 }
+
+export const DotVariant = {
+  available: "bg-green-200",
+  busy: "bg-orange-200",
+  full: "bg-red-200",
+};
+export type DotStatus = keyof typeof DotVariant;
